unrooted_must_root lint in seperate file
